requirements.txt
# 🔐 Crypto + Signing
cryptography==42.0.5
pyjwt[crypto]==2.8.0

# 🧪 Validation
pydantic==2.7.1

# 🧬 Environment management
python-dotenv==1.0.1

# ⚡ HTTP client for agents (e.g. Apple ID, APIs)
httpx==0.27.0

# 🛠️ CLI + scripting
argparse==1.4.0

# 🧪 Testing
pytest==8.2.2

# 📦 Optional OAuth / Apple ID agents
requests==2.32.3

Flask
Flask-CORS
python-dotenv

google-auth
google-auth-oauthlib
